+++ assets/themes/app.config.json
@@ -99,8 +99,8 @@
           },
           {
             "enabled": true,
-            "type": "encoding",
-            "titleL10n": "settings_ListViewTileTitle_encoding",
+            "type": "mediaSettings",
+            "titleL10n": "settings_ListViewTileTitle_mediaSettings",
             "icon": "0xf1cf"
           },
           {
