+++ lib/features/settings/features/media_settings/media_settings.dart
@@ -1,3 +1,4 @@
 export 'cubit/media_settings_cubit.dart';
+export 'cubit/media_settings_state.dart';
 export 'view/media_settings_screen.dart';
 export 'view/media_settings_screen_page.dart';
