+++ lib/features/settings/features/media_settings/view/media_settings_screen.dart
@@ -4,12 +4,16 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 
 import 'package:webtrit_phone/l10n/l10n.dart';
 import 'package:webtrit_phone/models/audio_processing_settings.dart';
-import 'package:webtrit_phone/models/enableble.dart';
 import 'package:webtrit_phone/models/encoding_settings.dart';
 import 'package:webtrit_phone/models/rtp_codec_profile.dart';
 import 'package:webtrit_phone/models/video_capturing_settings.dart';
 import 'package:webtrit_phone/widgets/widgets.dart';
 
+import 'components/choosable_section.dart';
+import 'components/heading_section.dart';
+import 'components/inline_choosable_section.dart';
+import 'components/reorderable_section.dart';
+import 'components/slidable_section.dart';
 import '../media_settings.dart';
 
 class MediaSettingsScreen extends StatefulWidget {
@@ -401,318 +405,6 @@ class _MediaSettingsScreenState extends State<MediaSettingsScreen> {
           selected: videoCapturingSettings.framerate,
           onSelect: (option) => cubit.setVideoCapturingSettings(videoCapturingSettings.copyWithFramerate(option)),
         ),
-      ],
-    );
-  }
-}
-
-class HeadingSection extends StatelessWidget {
-  const HeadingSection({
-    required this.title,
-    required this.tooltip,
-    required this.icon,
-    super.key,
-  });
-
-  final String title;
-  final String tooltip;
-  final Widget icon;
-
-  @override
-  Widget build(BuildContext context) {
-    final textStyle = Theme.of(context).textTheme.titleMedium!.copyWith(fontWeight: FontWeight.w600);
-
-    return Row(
-      children: [
-        icon,
-        const SizedBox(width: 8),
-        Expanded(child: Text(title, style: textStyle)),
-        const SizedBox(width: 8),
-        Tooltip(
-          message: tooltip,
-          triggerMode: TooltipTriggerMode.tap,
-          padding: const EdgeInsets.all(16),
-          margin: const EdgeInsets.all(16),
-          showDuration: const Duration(seconds: 10),
-          child: const Icon(Icons.info_outline),
-        )
-      ],
-    );
-  }
-}
-
-class ChoosableSection<T> extends StatelessWidget {
-  const ChoosableSection({
-    required this.title,
-    required this.buildOptionTitle,
-    required this.options,
-    required this.selected,
-    required this.onSelect,
-    super.key,
-  });
-
-  final String? title;
-  final Widget Function(T?) buildOptionTitle;
-
-  final List<T> options;
-  final T? selected;
-  final Function(T?) onSelect;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        if (title != null) ...[
-          Row(
-            children: [
-              const SizedBox(width: 4),
-              Expanded(child: Text(title!)),
-              const SizedBox(width: 4),
-            ],
-          ),
-          const SizedBox(height: 8.0),
-        ],
-        ListTile(
-          selected: selected == null,
-          title: buildOptionTitle(null),
-          onTap: () => onSelect(null),
-          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-          leading: Icon(selected == null ? Icons.radio_button_checked : Icons.radio_button_unchecked),
-          // contentPadding: EdgeInsets.zero,
-          visualDensity: VisualDensity.compact,
-          minTileHeight: 0,
-        ),
-        for (final option in options)
-          ListTile(
-            selected: selected == option,
-            title: buildOptionTitle(option),
-            onTap: () => onSelect(option),
-            shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-            leading: Icon(selected == option ? Icons.radio_button_checked : Icons.radio_button_unchecked),
-            // contentPadding: EdgeInsets.zero,
-            visualDensity: VisualDensity.compact,
-            minTileHeight: 0,
-          ),
-      ],
-    );
-  }
-}
-
-class InlineChoosableSection<T> extends StatelessWidget {
-  const InlineChoosableSection({
-    required this.title,
-    required this.buildOptionTitle,
-    required this.options,
-    required this.selected,
-    required this.onSelect,
-    super.key,
-  });
-
-  final String? title;
-  final Widget Function(T?) buildOptionTitle;
-
-  final List<T> options;
-  final T? selected;
-  final Function(T?) onSelect;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        if (title != null) ...[
-          Row(
-            children: [
-              const SizedBox(width: 4),
-              Expanded(child: Text(title!)),
-              const SizedBox(width: 4),
-              ToggleButtons(
-                borderRadius: BorderRadius.circular(12),
-                isSelected: [selected == null, ...options.map((option) => selected == option)],
-                children: [buildOptionTitle(null), for (final option in options) buildOptionTitle(option)],
-                onPressed: (index) {
-                  if (index == 0) {
-                    onSelect(null);
-                  } else {
-                    onSelect(options[index - 1]);
-                  }
-                },
-              )
-            ],
-          ),
-          // const SizedBox(height: 8.0),
-        ],
-      ],
-    );
-  }
-}
-
-class SlidableSection<T> extends StatelessWidget {
-  const SlidableSection({
-    required this.title,
-    required this.optionPrefix,
-    required this.buildOptionLabel,
-    required this.options,
-    required this.selected,
-    required this.onSelect,
-    super.key,
-  });
-
-  final String title;
-  final String optionPrefix;
-  final String Function(T?) buildOptionLabel;
-
-  final List<T> options;
-  final T? selected;
-  final Function(T?) onSelect;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Row(
-          children: [
-            const SizedBox(width: 4),
-            Expanded(child: Text(title)),
-            const SizedBox(width: 4),
-          ],
-        ),
-        Slider(
-          value: selected == null ? 0 : options.indexOf(selected as T) + 1,
-          secondaryTrackValue: 0.1,
-          onChanged: (v) {
-            if (v == 0) {
-              onSelect(null);
-            } else {
-              final index = v.round() - 1;
-              onSelect(options[index]);
-            }
-          },
-          min: 0,
-          max: options.length.toDouble(),
-          divisions: options.length + 1,
-          label: buildOptionLabel(selected),
-        ),
-        Row(
-          children: [
-            const SizedBox(width: 24),
-            Text(optionPrefix),
-            Text(
-              buildOptionLabel(selected),
-              style: TextStyle(
-                color: Theme.of(context).colorScheme.secondary,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            const SizedBox(width: 12),
-          ],
-        ),
-      ],
-    );
-  }
-}
-
-class ReorderableSection<T> extends StatelessWidget {
-  const ReorderableSection({
-    required this.title,
-    required this.warningTitle,
-    required this.warningMessage,
-    required this.buildOptionTitle,
-    required this.enabled,
-    required this.onEnable,
-    required this.items,
-    required this.onChange,
-    super.key,
-  });
-
-  final String title;
-  final String warningTitle;
-  final String warningMessage;
-  final Widget Function(T?) buildOptionTitle;
-
-  final bool enabled;
-  final Function(bool) onEnable;
-  final List<Enableble<T>> items;
-  final Function(List<Enableble<T>>) onChange;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Row(
-          children: [
-            const SizedBox(width: 4),
-            Expanded(child: Text(title)),
-            const SizedBox(width: 4),
-            Switch(value: enabled, onChanged: onEnable),
-            const SizedBox(width: 8.0),
-          ],
-        ),
-        if (enabled) ...[
-          Column(
-            children: [
-              Text(
-                warningTitle,
-                style: TextStyle(
-                  color: Theme.of(context).colorScheme.error,
-                  fontWeight: FontWeight.bold,
-                  fontSize: 13.0,
-                ),
-              ),
-              Text(
-                warningMessage,
-                style: TextStyle(
-                  color: Theme.of(context).colorScheme.error,
-                  fontSize: 12.0,
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 16),
-          ReorderableListView(
-            onReorder: (oldIndex, newIndex) {
-              if (oldIndex < newIndex) newIndex -= 1;
-
-              final newOptions = List<Enableble<T>>.from(items);
-              final profile = newOptions.removeAt(oldIndex);
-              newOptions.insert(newIndex, profile);
-
-              onChange(newOptions);
-            },
-            shrinkWrap: true,
-            buildDefaultDragHandles: false,
-            physics: const NeverScrollableScrollPhysics(),
-            children: [
-              for (final item in items)
-                ListTile(
-                  enabled: enabled,
-                  key: ValueKey(item),
-                  title: buildOptionTitle(item.option),
-                  leading: ReorderableDragStartListener(
-                    index: items.indexOf(item),
-                    enabled: enabled,
-                    child: const Icon(Icons.drag_handle),
-                  ),
-                  trailing: Checkbox(
-                    value: item.enabled,
-                    onChanged: enabled
-                        ? (v) {
-                            if (v == null) return;
-                            onChange(items.map((i) {
-                              if (i == item) {
-                                return (option: i.option, enabled: v);
-                              } else {
-                                return i;
-                              }
-                            }).toList());
-                          }
-                        : null,
-                  ),
-                  visualDensity: VisualDensity.compact,
-                  minTileHeight: 0,
-                ),
-            ],
-          ),
-        ],
       ],
     );
   }
